Refactor: Split monolithic server.js into modular architecture (AUD-01)
