+++ controllers/retentionController.js
@@ -1,0 +1,54 @@
+const { supabase } = require('../config');
+const { validateWalletAddress } = require('../middleware/verifyAdmin');
+
+const getRetentionPolicies = async (req, res) => {
+    try {
+        const { data: policies, error } = await supabase
+            .from('retention_policies')
+            .select('*')
+            .eq('is_active', true)
+            .order('created_at', { ascending: false });
+        if (error) throw error;
+        res.json({ success: true, policies });
+    } catch (error) {
+        console.error('Get retention policies error:', error);
+        res.status(500).json({ error: 'Failed to get retention policies' });
+    }
+};
+
+const createRetentionPolicy = async (req, res) => {
+    try {
+        const { name, caseType, retentionDays, archiveMethod, jurisdiction, lawReference, userWallet } = req.body;
+        if (!validateWalletAddress(userWallet)) {
+            return res.status(400).json({ error: 'Invalid wallet address' });
+        }
+        const { data: policy, error } = await supabase
+            .from('retention_policies')
+            .insert({
+                name, case_type: caseType, retention_days: retentionDays,
+                archive_method: archiveMethod, jurisdiction,
+                law_reference: lawReference, created_by: userWallet
+            })
+            .select().single();
+        if (error) throw error;
+        res.json({ success: true, policy });
+    } catch (error) {
+        console.error('Create retention policy error:', error);
+        res.status(500).json({ error: 'Failed to create retention policy' });
+    }
+};
+
+const exportTimelinePdf = async (req, res) => {
+    try {
+        const { caseId, evidence } = req.body;
+        const pdfContent = `CASE EVIDENCE TIMELINE REPORT\nCase ID: ${caseId}\nGenerated: ${new Date().toLocaleString()}\nTotal Evidence Items: ${evidence.length}\n\nEVIDENCE CHRONOLOGY:\n${evidence.map((item, i) => `${i+1}. ${item.title}\n   Type: ${item.type}\n   Date: ${new Date(item.timestamp).toLocaleString()}\n   Submitted by: ${item.submitted_by.substring(0,8)}...\n   Hash: ${item.hash}\n`).join('')}\nGenerated by EVID-DGC Blockchain Evidence Management System.`;
+        res.setHeader('Content-Type', 'application/pdf');
+        res.setHeader('Content-Disposition', `attachment; filename="timeline_${caseId}_${new Date().toISOString().split('T')[0]}.pdf"`);
+        res.send(Buffer.from(pdfContent));
+    } catch (error) {
+        console.error('Timeline PDF export error:', error);
+        res.status(500).json({ error: 'Failed to export timeline as PDF' });
+    }
+};
+
+module.exports = { getRetentionPolicies, createRetentionPolicy, exportTimelinePdf };
