+++ middleware/upload.js
@@ -1,0 +1,42 @@
+const multer = require('multer');
+
+// Configure multer for file uploads
+const upload = multer({
+    limits: {
+        fileSize: 100 * 1024 * 1024 // 100MB limit
+    },
+    fileFilter: (req, file, cb) => {
+        const allowedTypes = [
+            'application/pdf',
+            'image/jpeg',
+            'image/jpg', 
+            'image/png',
+            'image/gif',
+            'video/mp4',
+            'video/avi',
+            'video/mov',
+            'audio/mp3',
+            'audio/wav',
+            'audio/m4a',
+            'application/msword',
+            'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
+            'application/vnd.ms-excel',
+            'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
+            'text/plain',
+            'application/zip',
+            'application/x-rar-compressed'
+        ];
+        
+        if (allowedTypes.includes(file.mimetype)) {
+            cb(null, true);
+        } else {
+            cb(new Error(`File type ${file.mimetype} not supported`), false);
+        }
+    },
+    onError: (err, next) => {
+        console.error('Multer error:', err);
+        next(err);
+    }
+});
+
+module.exports = upload;
