+++ routes/healthRoutes.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+const { getHealth } = require('../controllers/healthController');
+
+router.get('/health', getHealth);
+
+module.exports = router;
